==> 2.scripts/1_script_simulating_data.Rmd <==
---
title: "Simulating Environmental Data on the Effect of Heat Waves on Years of Life Lost"
description: |
  Detailled Script.
author:
  - name: Tarik Benmarhnia
    url: https://profiles.ucsd.edu/tarik.benmarhnia
    affiliation: UCSD & Scripps Institute
    affiliation_url: https://benmarhniaresearch.ucsd.edu/
  - name: Marie-Abèle Bind 
    url: https://scholar.harvard.edu/marie-abele
    affiliation: Biostatistics Center, Massachusetts General Hospital
    affiliation_url: https://biostatistics.massgeneral.org/faculty/marie-abele-bind-phd/
  - name: Léo Zabrocki 
    url: https://lzabrocki.github.io/
    affiliation: Paris School of Economics
    affiliation_url: https://www.parisschoolofeconomics.eu/fr/zabrocki-leo/
date: "`r Sys.Date()`"
output: 
    distill::distill_article:
      keep_md: true
      toc: true
      toc_depth: 3
---

```{r setup, include=FALSE}
# code chunk option
knitr::opts_chunk$set(
  echo = TRUE,
  message = FALSE,
  warning = FALSE,
  layout="l-body-outset",
  dev = "CairoPNG",
  dpi = 600
)
```

<style>
body {
text-align: justify}
</style>

In this document, we provide all steps and R codes required to simulate the environmental data. For each day, we simulate the number of years of life lost (YoLL) with and without the occurrence of an heat wave. **Should you have any questions, need help to reproduce the simulation or find coding errors, please do not hesitate to contact us at leo.zabrocki@psemail.eu**

# Required Packages

To reproduce exactly the `1_script_simulating_data.html` document, we first need to have installed:

* the [R](https://www.r-project.org/) programming language 
* [RStudio](https://rstudio.com/), an integrated development environment for R, which will allow you to knit the `1_script_simulating_data.Rmd` file and interact with the R code chunks
* the [R Markdown](https://rmarkdown.rstudio.com/) package
* and the [Distill](https://rstudio.github.io/distill/) package which provides the template for this document. 

Once everything is set up, we load the following packages:

```{r}
# load required packages
library(knitr) # for creating the R Markdown document
library(here) # for files paths organization
library(tidyverse) # for data manipulation and visualization
library(ggridges) # for ridge density plots
library(Cairo) # for printing custom police of graphs
library(DT) # for displaying the data as tables
```

We finally load our custom `ggplot2` theme for graphs:

```{r}
# load ggplot custom theme
source(here::here("2.scripts",
                  "functions",
                  "script_theme_tufte.R"))
# define nice colors
my_blue <- "#0081a7"
my_orange <- "#fb8500"
```

# Environmental Data

We base our simulation on a dataset that includes all residents of the island of Montreal, Canada, who died from non-accidental causes in the city during the summer periods (June-August) between 1990 and 2007. Daily observations with missing values were dropped to ease the simulation procedure. We load the environmental dataset `raw_environmental_data.rds` located in the `1.data` folder:

```{r}
# load environmental data
data <-
  readRDS(here::here("1.data", "raw_environmental_data.rds"))
```

We display below five of the variables included in the data:

```{r}
# display part of the data
datatable(data %>% select(id, date, yll, total_death, heat_wave))
```


The dataset contains `r nrow(data)` daily observations. We provide below the definition of the variables. If a variable's name includes the term "lag", it indicates the lag of this variable (e.g. `o3_lag_1` represents the first daily lag of the ozone concentration). 

* `id`: the daily unit index in the data.
* `date`: the date of the observation in year-month-day format.
* `weekday`: the day of the week.
* `week`: the week of the year.
* `month`: the month of the year.
* `year`: the year.
* `yll`: total daily years of life lost.
* `total_death`: daily number of non-accidental deaths.
* `heat_wave`: dummy equal to 1 if a heat wave occurred on that day and to 0 otherwise.
* `humidity_relative`: relative humidity in %.
* `temperature_average`: average temperature in °C.
* `temperature_maximum`: maximum temperature in °C.
* `temperature_minimum`: minimum temperature in °C.
* `o3`: daily average concentration of ozone ($\mu g/m^3$).
* `no2`: daily average concentration of nitrogen dioxide ($\mu g/m^3$).
* The rest of the variables are dummy indicators for the week of the day, the week of the year, the month and the year.


We quickly explore the raw data. We first plot below the distribution of heat waves over time:

```{r, fig.width=8, fig.height=5, code_folding="Please show me the code!"}
# make stripes graph
graph_stripes_treatment <- data %>%
  mutate(is_treated = ifelse(heat_wave == 1, "Treated", "Control")) %>%
  ggplot(., aes(x = date, y = 1, fill = is_treated)) +
  geom_tile() +
  scale_y_continuous(expand = c(0, 0)) +
  facet_wrap( ~ year, scales = "free") +
  scale_fill_manual(name = "Daily Observations:", values = c(my_blue, my_orange)) +
  xlab("Date") + ylab("") +
  theme_tufte() +
  theme(axis.ticks.y = element_blank(),
        axis.text.y = element_blank())

# display the graph
graph_stripes_treatment

# save the graph
ggsave(
  graph_stripes_treatment,
  filename = here::here("3.outputs", "2.graphs", "graph_stripes_treatment.pdf"),
  width = 30,
  height = 15,
  units = "cm",
  device = cairo_pdf
)
```

We then plot the density distribution of variables according to the heat wave status:

```{r, fig.width=10, fig.height=5, code_folding="Please show me the code!"}
# create the graph
graph_cov_densities <- data %>%
  pivot_longer(
    cols = c(
      humidity_relative,
      temperature_minimum,
      temperature_average,
      temperature_maximum,
      o3,
      no2
    ),
    names_to = "variable",
    values_to = "value"
  ) %>%
  mutate(
    variable = case_when(
      variable == "humidity_relative" ~ "Relative Humidity (%)",
      variable == "temperature_minimum" ~ "Minimum Temperature (°C)",
      variable == "temperature_average" ~ "Average Temperature (°C)",
      variable == "temperature_maximum" ~ "Maxmimum Temperature (°C)",
      variable == "o3" ~ "O3 (µg/m³)",
      variable == "no2" ~ "NO2 (µg/m³)"
    )
  ) %>%
  mutate(
    variable = fct_relevel(
      variable,
      "Minimum Temperature (°C)",
      "Average Temperature (°C)",
      "Maxmimum Temperature (°C)",
      "Relative Humidity (%)",
      "NO2 (µg/m³)",
      "O3 (µg/m³)"
    )
  ) %>%
  mutate(heat_wave = ifelse(heat_wave == 1, "Days with Heat Wave", "Days without Heat Wave")) %>%
  ggplot(., aes(x = value, colour = fct_rev(heat_wave))) +
  geom_density() +
  scale_color_manual(name = "Group:", values = c(my_blue, my_orange)) +
  facet_wrap(~ variable, scales = "free") +
  ggtitle("Density Distribution of Weather and Air Pollution Variables") +
  xlab("Value") + ylab("Density") +
  theme_tufte() +
  theme(axis.ticks.y = element_blank(),
        axis.text.y = element_blank())

# display the graph
graph_cov_densities

# save the graph
ggsave(
  graph_cov_densities + labs(title = NULL),
  filename = here::here("3.outputs", "2.graphs", "graph_cov_densities.pdf"),
  width = 20,
  height = 10,
  units = "cm",
  device = cairo_pdf
)
```


# Simulating Potential Outcomes

For each day, we simulate the number of years of life lost with and without the occurrence of a heat wave. Following the Neyman-Rubin causal model:

* $\text{W}_{i}$ denotes the heat wave indicator on a certain day $i$ and $\text{Y}_{i}$ is the outcome of interest (i.e., the daily number of years of life loss).
* $\text{Y}_{i}(\text{W}_{i}=0)$ and $\text{Y}_{i}(\text{W}_{i}=1)$ are respectively the potential outcomes had a heat wave not occurred and occurred.
* The observed outcome is defined as $\text{Y}_{i}^{\text{obs}} = \text{W}_{i} \times \text{Y}_{i}(\text{W}_{i}=1) + (1-\text{W}_{i}) \times \text{Y}_{i}(\text{W}_{i}=0)$

In this simulation exercise, we are interested in measuring the average treatment effect on the treated (ATT), that is to say the average effect of heatwave on the daily years of life lost for days that were actually exposed: 

\begin{equation}
\text{ATT} = \sum_{i:\text{W}_{i}=1}^{\text{N}_{\text{T}}}\text{Y}(\text{W}_{i}=1) - \text{Y}(\text{W}_{i}=0)
\end{equation}

where $\text{N}_{\text{T}}$ is the number of days in the data where a heatwave occured. 

We simulated the potential outcomes $\text{Y}_{i}(\text{W}_{i}=0)$ and $\text{Y}_{i}(\text{W}_{i}=1)$  (i=1,..., N=1,376) using two normal random variables with variance equal to 1 and respective means $\mu_{0i}$ and $\mu_{1i}$ such that:

* $\mu_{0i} = \alpha_{0} + \sum_{k=1}^{3}\alpha_{\text{HW}lag_k}\text{HW}_{lag_k} + \sum_{k=1}^{3}\alpha_{\text{O}_{3lag_{k}}}\text{O}_{3lag_{k}}+ \\ \sum_{k=1}^{3}\alpha_{\text{NO}_{2lag_{k}}}\text{NO}_{2lag_{k}} + \alpha_{\text{Humidity}}\text{Humidity} + \alpha_{\text{July}}\text{July} + \alpha_{\text{August}}\text{August} + \\ \sum_{k=\{M, Tu, Th, F, Sa, Su}\alpha_{k}\text{DoW}_{k} + \sum_{k=23}^{35}\alpha_{\text{week}_{k}}\text{week}_{k} + \sum_{k=1991}^{2007}\alpha_{\text{year}_{k}}\text{year}_{k}$ 

* $\mu_{1i} = \alpha_{0} + \boldsymbol{\alpha_{\text{HW}lag_{0}}\text{HW}_{lag_{0}}} + \sum_{k=1}^{3}\alpha_{\text{HW}lag_k}\text{HW}_{lag_k} + \sum_{k=1}^{3}\alpha_{\text{O}_{3lag_{k}}}\text{O}_{3lag_{k}}+ \\ \sum_{k=1}^{3}\alpha_{\text{NO}_{2lag_{k}}}\text{NO}_{2lag_{k}} + \alpha_{\text{Humidity}}\text{Humidity} + \alpha_{\text{July}}\text{July} + \alpha_{\text{August}}\text{August} + \\ \sum_{k=\{M, Tu, Th, F, Sa, Su}\alpha_{k}\text{DoW}_{k} + \sum_{k=23}^{35}\alpha_{\text{week}_{k}}\text{week}_{k} + \sum_{k=1991}^{2007}\alpha_{\text{year}_{k}}\text{year}_{k}$

where $\alpha_{0}=2000, \alpha_{HWlag_{0}} = 230, \alpha_{HWlag_{1}} = 180, \\\alpha_{HWlag_{2}} = 180, \alpha_{O_{3lag_{1}}} = 3, \alpha_{O_{3lag_{2}}} = \alpha_{O_{3lag_{3}}} = 0, \\ \alpha_{NO_{2lag_{1}}} = \alpha_{NO_{2lag_{2}}} = \alpha_{NO_{2lag_{3}}} = 0, \\ \alpha_{July} = \alpha_{August} = 0, \\\alpha_{Monday} = 280, \alpha_{Tuesday} = 292, \alpha_{Thursday} = 290, \alpha_{Friday} = 365, \alpha_{Saturday} = 280, \\\alpha_{Sunday} = 315, \alpha_{week_{23}} = \alpha_{week_{24}} = \alpha_{week_{25}} = \alpha_{week_{26}} = \\ \alpha_{week_{27}} = \alpha_{week_{28}} = \alpha_{week_{29}} = \alpha_{week_{30}} = \alpha_{week_{31}} = \\ \alpha_{week_{32}} = \alpha_{week_{33}} = \alpha_{week_{34}} = \alpha_{week_{35}} = 0, \alpha_{year_{1991}} = \\ \alpha_{year_{1992}} = \alpha_{year_{1993}} = \alpha_{year_{1994}} = \alpha_{year_{1995}} = \alpha_{year_{1996}} = \alpha_{year_{1997}} = \alpha_{year_{1998}} = \alpha_{year_{1998}} = 0, \\\alpha_{year_{1999}} = 240, \alpha_{year_{2000}} = 340, \alpha_{year_{2001}} = 450, \alpha_{year_{2002}} = 450, \alpha_{year_{2003}} = 0, \\ \alpha_{year_{2004}} = 0, \alpha_{year_{2005}} = 0, \alpha_{year_{2005}} = 210,  \alpha_{year_{2006}} = 0 \ \text{and} \ \alpha_{year_{2007}} = 0$.

The values of the variables' coefficients are motivated by the estimated coefficients when we regress the daily years of life lost on heat wave indicators, ozone and nitrogen dioxide concentrations, relative humidity and time indicators:

```{r}
# we fit a simple time stratified model
lm(
  yll ~ heat_wave + heat_wave_lag_1 + heat_wave_lag_2 + heat_wave_lag_3 +
    o3 + o3_lag_1 + o3_lag_2 + o3_lag_3 +
    no2 + no2_lag_1 + no2_lag_2 + no2_lag_3 +
    humidity_relative + relevel(as.factor(weekday), ref = "wednesday") +
    relevel(as.factor(week), ref = "22") + relevel(as.factor(month), ref = "june") +
    relevel(as.factor(year), ref = "1990"),
  data = data
) %>%
  # format output
  broom::tidy(., conf.int = TRUE) %>%
  # select relevant variables
  select(term, estimate, conf.low, conf.high) %>%
  # round estimates
  mutate_at(vars(estimate:conf.high), ~ round(., 1)) %>%
  # rename variables
  rename(
    "Term" = "term",
    "Estimate" = estimate,
    "Lower Bound of 95% CI" = conf.low,
    "Upper Bound of 95% CI" = conf.high
  ) %>%
  # display results
  datatable(.,  options = list(columnDefs = list(
    list(className = 'dt-center', targets = 2:4)
  )))
```

The potential outcomes are then constructed using the following code:

```{r}
# first set values of coefficients
alpha_0 <- 2000
alpha_hw <- 230
alpha_hw_lag_1 <- 180
alpha_hw_lag_2 <- 180
alpha_no2 <- 6
alpha_o3_lag_1 <- 3
alpha_year_1999 <- 240
alpha_year_2000 <- 340
alpha_year_2001 <- 340
alpha_year_2002 <- 450
alpha_year_2005 <- 210
alpha_friday <- 365
alpha_monday <- 280
alpha_saturday <- 280
alpha_sunday <- 315
alpha_thursday <- 290
alpha_tuesday <- 292

# then compute the two potential outcomes
# for each daily unit
data <- data %>%
  rowwise() %>%
  mutate(
    mu_0_i = alpha_0 + alpha_hw_lag_1 * heat_wave_lag_1 + alpha_hw_lag_2 * alpha_hw_lag_2 +
      alpha_no2 * no2 + alpha_o3_lag_1 * o3_lag_1 +
      alpha_monday * weekday_monday + alpha_tuesday * weekday_tuesday + 
      alpha_thursday * weekday_thursday + alpha_friday * weekday_friday + 
      alpha_saturday * weekday_saturday + alpha_sunday * weekday_sunday +
      alpha_year_1999 * year_1999 +
      alpha_year_2000 * year_2000 + alpha_year_2001 * year_2001 + 
      alpha_year_2002 * year_2002 + alpha_year_2005 + year_2005,
    mu_1_i = mu_0_i + (alpha_hw * heat_wave),
    y_0 = rnorm(1, mean = mu_0_i, sd = 1),
    y_1 = rnorm(1, mean = mu_1_i, sd = 1)
  ) %>%
  ungroup()
```

The average effect for heatwave for treated days is set to $+230$ daily years of life lost ( $\alpha_{HWlag_{0}}=230$). We can check that our simulation leads to the correct figure by computing, for treated units, the average difference between the $Y_{i}(1)$ and $Y_{i}(0)$:

```{r}
data %>%
  # select days where an heat wave occurred
  filter(heat_wave == 1) %>%
  # compute the average difference in potential outcomes
  summarise(ATT = mean(y_1 - y_0) %>% round(., 0))
```

We can assess this difference by plotting the distribution of the potential outcomes for treated units:

```{r, fig.width=8, fig.height=4, code_folding="Please show me the code!"}
# compute the average of potential outcomes
data_mean_po <- data %>%
  # select days where an heat wave occurred
  filter(heat_wave == 1) %>%
  # reshape potential outcomes into long format
  rename("Y(0)" = y_0, "Y(1)" = y_1) %>%
  pivot_longer(
    cols = c("Y(0)", "Y(1)"),
    names_to = "potential_outcomes",
    values_to = "values"
  ) %>% # compute means
  group_by(potential_outcomes) %>%
  summarise(mean_po = mean(values))

# make the graph
graph_po_densities <- data %>%
  # select days where an heat wave occurred
  filter(heat_wave == 1) %>%
  # reshape potential outcomes into long format
  rename("Y(0)" = y_0, "Y(1)" = y_1) %>%
  pivot_longer(
    cols = c("Y(0)", "Y(1)"),
    names_to = "potential_outcomes",
    values_to = "values"
  ) %>%
  # plot the density of potential outcomes
  ggplot(., aes(x = values, color = potential_outcomes)) +
  geom_density(key_glyph = "polygon") +
  geom_vline(
    data = data_mean_po,
    aes(xintercept = mean_po, colour = potential_outcomes),
    key_glyph = "blank"
  ) +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 10)) +
  scale_color_manual(name = "Potential Outcomes:", values = c(my_blue, my_orange)) +
  ggtitle("Distribution of Potential Outcomes by Treatment Status") +
  xlab("Daily Number of Years of Life Lost") + ylab("Density") +
  theme_tufte() +
  theme(axis.ticks.y = element_blank(),
        axis.text.y = element_blank())

# display the graph
graph_po_densities

# save the graph
ggsave(
  graph_po_densities + labs(title = NULL),
  filename = here::here("3.outputs", "2.graphs", "graph_po_densities.pdf"),
  width = 15,
  height = 8,
  units = "cm",
  device = cairo_pdf
)
```
We can also compute the average treatment effect for all units (ATE):

```{r}
data %>%
  # compute the average difference in potential outcomes
  summarise(ATE = mean(y_1 - y_0) %>% round(., 0))
```

For each unit, we finally express the observed number of years of life lost according to the treatment assignment:

```{r}
# create observed years of life lost
data <- data %>%
  mutate(y_obs = ifelse(heat_wave == 1, y_1, y_0))
```

We reorder and save the data:

```{r}
# reorder and save the data
data %>%
  select(id, date:year, y_obs, y_1, y_0, heat_wave:year_2007) %>%
  saveRDS(., here::here("1.data", "simulated_environmental_data.rds"))
```
